fix: end-of-frame detection.
Do not suppose end-of-frame is aligned with storage's end-of-file.
Handy when zstd frames are part of larger streams, used to return
some puzzling "Context should be init first" error.

Zstd natively recognizes frame termination and signal the situation
through the special `0` value of the returned variant:
"ZBUFF_decompressContinue returns either a hint to preferred nb
of bytes to use as input for next function call or 0 when a frame
is completely decoded, or an error code, which can be tested using
ZBUFF_isError()".

Closes #3.