`Decoder` seemlessly concatenate compressed frames
Can be disabled with `Decoder::single_frame()`
Added a `test_concatenated_frames` for that.

`Decoder::with_dictionary` now takes a `Into<Vec<u8>>`, as we need to
keep the dictionary around to re-initialize the context on concatenated
frames.
To avoid cloning the dictionary for each Decoder, a
`with_shared_dictionary` taking an `Arc<Vec<u8>>` is also added.