encoder: add try_finish to allow working with non-blocking IO
With non-blocking IO, `finish()` is insufficient: it might error out with
`WouldBlock`, in which case the solution is to retry next time, not just error
out. To do that, we need a variant of `finish()` that doesn't consume itself on
error. That's precisely what `try_finish` does.

The added tests make sure that writes after a failed `try_finish` panic, and
that internal state is appropriately updated after each state of a
`try_finish`.