encoder: loop to make sure successful results are returned
Consider what happens if the first few attempts at writing succeed but then the
write fails with a non-`Interrupted` error. In that case, at the moment we
bubble up the error. This is wrong since writing out the input into the zstd
compressor did succeed, and the compressor accepted the input. Retrying the
input would be wrong.

With this change, we treat a write to the zstd compressor buffer as a point
from which a positive result can correctly be returned.