fix: Set ZSTD_MULTITHREAD when built with zstdmt
Advanced features like multithreading are only available when libzstd
is statically linked per documentation. pkg-config is not very good
for static linking. That means users have to rely on this crate to build it
correctly.

Current `build.rs` is not passing `ZSTD_MULTITHREAD` flag to cc and
because of that resulting library built without multithreading support.

Resolves: #64