zstd-sys: drop dependency on itertools
Itertools::join is only marginally more convenient than using the
built-in slice join method. Itertools::join is slightly more efficient,
but that's not really a concern in a build script.

So, all things considered, I think it's worth avoid the itertools
dependency in a low-level utility like this, for the benefit of
downstream consumers.