zstd-sys: avoid dependencies on itertools and glob (#103)
* zstd-sys: drop dependency on itertools

Itertools::join is only marginally more convenient than using the
built-in slice join method. Itertools::join is slightly more efficient,
but that's not really a concern in a build script.

So, all things considered, I think it's worth avoid the itertools
dependency in a low-level utility like this, for the benefit of
downstream consumers.

* zstd-sys: drop dependency on glob

Similar to the last commit, I think it's worth accepting a very slight
increase in complexity in this crate to avoid the dependency on the glob
crate.