Fix the `Decoder` API sometimes not outputting all available data
This is done by first asking zstd for more data, before trying to read
more compressed bytes from the data source.

An alternative would be the
[`BufReader::buffer`](https://doc.rust-lang.org/std/io/struct.BufReader.html#method.buffer)
function, but that is not available because we only have a type that
satisfies `BufRead` in this context.

I tested this with my TCP connection and it solved my problem.

Fixes #119.

I haven't tested this for performance regressions.