wasm32-unknown-unknown: Fix undefined malloc/free/calloc symbols (#275)
As discussed in https://github.com/gyscos/zstd-rs/issues/250, in some
cases the definitions of `malloc`, `free`, and `calloc` are used but
don't get linked in, leading to undefined symbols (i.e. reference to the
symbol in "env").

I was not able to determine exactly why this is happening, but it seems
to be related to `malloc`, `calloc`, and `free` being defined as inline
functions.  When building in debug mode there are undefined symbols;
when building in release mode there are not, presumably because the
functions are inlined.

Switching to macros avoids issues of inlining and fixes this issue for
me.