Add cfg attribute for advanced seekable impls and make seek table creation safe (#328)
* Fix missing cfg attribute on AdvancedSeekable impls

The missing cfg macros on the impls caused a compilation error when
compiling with feature `seekable` but without `std`

* Make seek table creation safe

The `SeekTable::try_from_seekable()` fn segfaulted when called with an
uninitialized seekable in zstd versions prior to v1.5.7. This is fixed
with v1.5.7, so the fn is safe to use now.